% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{dotplot}
\alias{dotplot}
\title{FUNCTION_TITLE}
\usage{
dotplot(data, x, y, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_segment}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{margin}}
\code{\link[stats]{reorder.default}}
\code{\link[ggthemes]{theme_fivethirtyeight}}
}
