% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_frames.R
\name{gpluck_get_index_scores}
\alias{gpluck_get_index_scores}
\title{FUNCTION_TITLE}
\usage{
gpluck_get_index_scores(patient)
}
\arguments{
\item{patient}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[here]{here}}
\code{\link[janitor]{clean_names}}
\code{\link[bwu]{gpluck_make_columns}}, \code{\link[bwu]{gpluck_make_score_ranges}}
\code{\link[tidytable]{mutate}}, \code{\link[tidytable]{case_when}}
\code{\link[glue]{glue}}
\code{\link[dplyr]{relocate}}, \code{\link[dplyr]{filter}}
\code{\link[readr]{write_delim}}
}
