% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.research.compendium.root-package.R
\docType{package}
\name{pj.research.compendium.root-package}
\alias{pj.research.compendium.root}
\alias{pj.research.compendium.root-package}
\title{pj.research.compendium.root: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brainworkup/pj.research.compendium.root}
  \item Report bugs at \url{https://github.com/brainworkup/pj.research.compendium.root/issues}
}

}
\author{
\strong{Maintainer}: Joey Trampush \email{j.trampush@gmail.com} (\href{https://orcid.org/0000-0003-0830-508X}{ORCID}) [copyright holder]

}
\keyword{internal}
